Deployment/Azure_CloudFunction/requirements.txt
azure-functions
requests
opencv-python
numpy
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
torchvision==0.6.0+cpu